Simplify parsing of default gamepad mappings
The outer glfwUpdateGamepadMappings function is now bypassed, as the
data in mappings.h does not contain any comments or line breaks.

This is done before joystick support has been initialized, so there is
no need to look for matching devices.

Finally, the array of default mappings is pre-allocated.  This has no
measurable performance impact but does generate a lot of calls, which
won't be nice for a user provided custom allocator to deal with.